refactor: 데이터 모델 및 모델 매퍼 수정
