+++ app/src/main/java/com/masonk/daehakjumak/presentation/model/MenuModel.kt
@@ -1,6 +1,6 @@
 package com.masonk.daehakjumak.presentation.model
 
-import com.masonk.daehakjumak.core.MenuType
+import com.masonk.daehakjumak.core.enums.MenuType
 
 data class MenuModel(
     val id: String, // 메뉴 id
